DocumentExtractionAgent/requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-mongodb
langchain-groq
pymongo
sentence-transformers
python-dotenv
pypdf
